Move integrators and importers to their own module
